# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Community {
  id: number;
  name: string;
  slug: string;
  color: string;
  memberCount: number;
  description: string;
}

export interface Author {
  id: number;
  displayName: string;
  username: string;
  avatar: string;
  communityName: string;
  communityColor: string;
  isLive: boolean;
  isPremium?: boolean;
}

export type PostType = typeof PostType[keyof typeof PostType];


export const PostType = {
  text: 'text',
  image: 'image',
  video: 'video',
} as const;

export interface Post {
  id: number;
  author: Author;
  type: PostType;
  text: string;
  caption: string;
  mediaUrl: string;
  link?: string | null;
  createdAt: string;
  likes: number;
  comments: number;
  shares: number;
  likedByViewer: boolean;
}

export interface FeedResponse {
  posts: Post[];
  viewer: Author;
}

export interface DiscoverResponse {
  communities: Community[];
  people: Author[];
  topics: string[];
}

export interface CommunityDetail {
  community: Community;
  members: Author[];
  posts: Post[];
}

export type CreatePostRequestType = typeof CreatePostRequestType[keyof typeof CreatePostRequestType];


export const CreatePostRequestType = {
  text: 'text',
  image: 'image',
  video: 'video',
} as const;

export interface CreatePostRequest {
  authorId: number;
  type: CreatePostRequestType;
  text?: string;
  caption?: string;
  mediaUrl?: string;
  link?: string;
}

export interface LikeRequest {
  viewerId: number;
}

export interface LikeResponse {
  liked: boolean;
  likes: number;
}

export type MessageStatus = typeof MessageStatus[keyof typeof MessageStatus];


export const MessageStatus = {
  sent: 'sent',
  delivered: 'delivered',
  read: 'read',
} as const;

export interface Message {
  id: number;
  senderId: number;
  recipientId: number;
  body: string;
  createdAt: string;
  status: MessageStatus;
}

export interface Conversation {
  id: number;
  person: Author;
  lastMessage: string;
  unreadCount: number;
  updatedAt: string;
}

export interface MessagesResponse {
  conversations: Conversation[];
}

export interface SendMessageRequest {
  senderId: number;
  recipientId: number;
  body: string;
}

export interface Profile {
  user: Author;
  followerCount: number;
  followingCount: number;
  posts: Post[];
}

export type AppNotificationKind = typeof AppNotificationKind[keyof typeof AppNotificationKind];


export const AppNotificationKind = {
  like: 'like',
  comment: 'comment',
  follow: 'follow',
  live: 'live',
  message: 'message',
  community: 'community',
} as const;

export interface AppNotification {
  id: number;
  kind: AppNotificationKind;
  title: string;
  detail: string;
  createdAt: string;
  read: boolean;
}

export interface NotificationsResponse {
  notifications: AppNotification[];
}

export type GetFeedParams = {
viewerId?: number;
};

export type GetDiscoverParams = {
query?: string;
};

export type GetMessagesParams = {
viewerId?: number;
};