# lib/api-zod/src/generated/types/appNotification.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AppNotificationKind } from './appNotificationKind';

export interface AppNotification {
  id: number;
  kind: AppNotificationKind;
  title: string;
  detail: string;
  createdAt: string;
  read: boolean;
}