# lib/api-zod/src/generated/types/community.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Community {
  id: number;
  name: string;
  slug: string;
  color: string;
  memberCount: number;
  description: string;
}