# lib/api-zod/src/generated/types/communityDetail.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Author } from './author';
import type { Community } from './community';
import type { Post } from './post';

export interface CommunityDetail {
  community: Community;
  members: Author[];
  posts: Post[];
}