# lib/api-zod/src/generated/types/createPostRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreatePostRequestType } from './createPostRequestType';

export interface CreatePostRequest {
  authorId: number;
  type: CreatePostRequestType;
  text?: string;
  caption?: string;
  mediaUrl?: string;
  link?: string;
}