# lib/api-zod/src/generated/types/discoverResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Author } from './author';
import type { Community } from './community';

export interface DiscoverResponse {
  communities: Community[];
  people: Author[];
  topics: string[];
}