# lib/api-zod/src/generated/types/feedResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Author } from './author';
import type { Post } from './post';

export interface FeedResponse {
  posts: Post[];
  viewer: Author;
}