# lib/api-zod/src/generated/types/messagesResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Conversation } from './conversation';

export interface MessagesResponse {
  conversations: Conversation[];
}