# lib/api-zod/src/generated/types/notificationsResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AppNotification } from './appNotification';

export interface NotificationsResponse {
  notifications: AppNotification[];
}