# lib/api-zod/src/generated/types/post.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Author } from './author';
import type { PostType } from './postType';

export interface Post {
  id: number;
  author: Author;
  type: PostType;
  text: string;
  caption: string;
  mediaUrl: string;
  link?: string | null;
  createdAt: string;
  likes: number;
  comments: number;
  shares: number;
  likedByViewer: boolean;
}