# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Author } from './author';
import type { Post } from './post';

export interface Profile {
  user: Author;
  followerCount: number;
  followingCount: number;
  posts: Post[];
}